+++ internal/clientset/clientset.go
@@ -35,6 +35,7 @@ import (
 	e2ms "github.com/coralogix/coralogix-management-sdk/go/openapi/gen/events2metrics_service"
 
 	globalRouters "github.com/coralogix/coralogix-management-sdk/go/openapi/gen/global_routers_service"
+	identity "github.com/coralogix/coralogix-management-sdk/go/openapi/gen/identity_service"
 	integrations "github.com/coralogix/coralogix-management-sdk/go/openapi/gen/integration_service"
 	ipaccess "github.com/coralogix/coralogix-management-sdk/go/openapi/gen/ip_access_service"
 	ams "github.com/coralogix/coralogix-management-sdk/go/openapi/gen/metrics_data_archive_service"
@@ -50,6 +51,7 @@ import (
 	archiveLogs "github.com/coralogix/coralogix-management-sdk/go/openapi/gen/target_service"
 	teamGroupss "github.com/coralogix/coralogix-management-sdk/go/openapi/gen/team_groups_management_service"
 	teamsservice "github.com/coralogix/coralogix-management-sdk/go/openapi/gen/teams_service"
+	usersservice "github.com/coralogix/coralogix-management-sdk/go/openapi/gen/users_management_service"
 
 	slos "github.com/coralogix/coralogix-management-sdk/go/openapi/gen/slos_service"
 )
@@ -59,7 +61,6 @@ type ClientSet struct {
 	dataSet        *cxsdk.DataSetClient
 	legacySlos     *cxsdk.LegacySLOsClient
 	ruleGroups     *cxsdk.RuleGroupsClient
-	users          *UsersClient
 	events2Metrics *e2ms.Events2MetricsServiceAPIService
 
 	dahboardsFolders      *dbfs.DashboardFoldersServiceAPIService
@@ -92,6 +93,9 @@ type ClientSet struct {
 	groups                *GroupsClient
 	teamGroups            *teamGroupss.TeamGroupsManagementServiceAPIService
 	teams                 *teamsservice.TeamsServiceAPIService
+	users                 *usersservice.UsersManagementServiceAPIService
+	identity              *identity.IdentityServiceAPIService
+	teamID                teamIDCache
 }
 
 func (c *ClientSet) ParsingRuleGroups() *prgs.RuleGroupsServiceAPIService {
@@ -190,8 +194,12 @@ func (c *ClientSet) Groups() *GroupsClient {
 	return c.groups
 }
 
-func (c *ClientSet) Users() *UsersClient {
+func (c *ClientSet) Users() *usersservice.UsersManagementServiceAPIService {
 	return c.users
+}
+
+func (c *ClientSet) Identity() *identity.IdentityServiceAPIService {
+	return c.identity
 }
 
 func (c *ClientSet) Scopes() *scopess.ScopesServiceAPIService {
@@ -258,8 +266,6 @@ func NewClientSet(region string, apiKey string, grpcTarget string) *ClientSet {
 		legacySlos:  cxsdk.NewLegacySLOsClient(grpcCreator),
 		ruleGroups:  cxsdk.NewRuleGroupsClient(grpcCreator),
 
-		users: NewUsersClient(region, apiKey),
-
 		events2Metrics: cs.Events2Metrics(),
 
 		dahboardsFolders:      cs.DashboardFolders(),
@@ -292,5 +298,7 @@ func NewClientSet(region string, apiKey string, grpcTarget string) *ClientSet {
 		groups:                NewGroupsClient(region, apiKey),
 		teamGroups:            cs.Groups(),
 		teams:                 cs.Teams(),
+		users:                 cs.Users(),
+		identity:              cs.Identity(),
 	}
 }